Authorization, Pagination added

+++ src/todo/todo.service.ts
@@ -2,6 +2,7 @@ import { BadRequestException, Injectable, NotFoundException } from '@nestjs/comm
 import { InjectModel } from '@nestjs/mongoose';
 import { Todo } from './schemas/todo.schema';
 import mongoose from 'mongoose';
+import { Query } from 'express-serve-static-core';
 
 @Injectable()
 export class TodoService {
@@ -11,8 +12,16 @@ export class TodoService {
         private toodoModel: mongoose.Model<Todo>
     ){}
     
-    async Findall():Promise<Todo[]>{
-        const todos = await this.toodoModel.find()
+    async Findall(query:Query):Promise<Todo[]>{
+
+        const resPerPage = 3;
+        const currentPage = Number(query.page) || 1;
+        const skip = resPerPage*(currentPage-1);
+
+        const todos = await this.toodoModel
+        .find()
+        .limit(resPerPage)
+        .skip(skip)
         return todos;
     }
 
@@ -39,6 +48,19 @@ export class TodoService {
 
     async updatetodo(id:string,todo:Todo):Promise<Todo>{
 
+        const isValidId = mongoose.isValidObjectId(id);
+        if(!isValidId)
+        {
+            throw new BadRequestException('Please enter correct id');
+        }
+
+        const _todo = await this.toodoModel.findById(id);
+        if(!_todo)
+        {
+            throw new NotFoundException('todo not found')
+        }
+    
+
         return await this.toodoModel.findByIdAndUpdate(id,todo,{
             new:true,
             runValidators:true
@@ -46,6 +68,19 @@ export class TodoService {
     }
     
     async deleteByid(id:string):Promise<Todo>{
+
+        const isValidId = mongoose.isValidObjectId(id);
+        if(!isValidId)
+            {
+                throw new BadRequestException('Please enter correct id');
+            }
+
+         const todo = await this.toodoModel.findById(id);
+        if(!todo)
+        {
+            throw new NotFoundException('todo not found')
+        }
+
        return await this.toodoModel.findByIdAndDelete(id);
        
     }
